backend/requirements.txt
fastapi==0.111.0
SQLAlchemy==2.0.30
pydantic==2.7.1
loguru==0.7.2
python-dotenv==1.0.1
psycopg2-binary # for postgres